♻️ moved Code from `SlicingSimulator` directly to `HybridSchrodingerFeynmanSimulator`
